All in one Translator
This project of developing TRANSLATOR is a console application by using the technology of PYTHON and using its different libraries. In this project, one can translate desired content into desired language and there will be voice-to-text between the user and the server and using this approach further it will translate the text in desired language and also there will be image translator also which will extract text from image and using this approach only further it will translate the text in desired language.